requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
-e git+https://github.com/UiL-OTS-labs/python-docx2txt#egg=docx2txt
    # via -r requirements.in
-e git+https://github.com/UiL-OTS-labs/django-shared-core.git@v2.0.10#egg=uil-django-core
    # via -r requirements.in
alabaster==0.7.16
    # via sphinx
arabic-reshaper==3.0.0
    # via xhtml2pdf
asgiref==3.7.2
    # via django
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
babel==2.14.0
    # via sphinx
black==23.12.1
    # via -r requirements.in
blessed==1.20.0
    # via curtsies
bpython==0.24
    # via -r requirements.in
cdh-django-core @ git+https://github.com/CentreForDigitalHumanities/django-shared-core.git@v3.1.0
    # via -r requirements.in
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via reportlab
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   djlint
    #   pyhanko
closure==20191111
    # via uil-django-core
colorama==0.4.6
    # via djlint
cryptography==41.0.7
    # via
    #   pyhanko
    #   pyhanko-certvalidator
    #   pyopenssl
    #   pysaml2
cssbeautifier==1.14.11
    # via djlint
cssselect2==0.7.0
    # via svglib
curtsies==0.4.2
    # via bpython
cwcwidth==0.1.9
    # via
    #   bpython
    #   curtsies
defusedxml==0.7.1
    # via
    #   djangosaml2
    #   pysaml2
deprecated==1.2.14
    # via cdh-django-core
django==3.2.23
    # via
    #   -r requirements.in
    #   cdh-django-core
    #   django-auth-ldap
    #   django-braces
    #   django-debug-toolbar
    #   django-extensions
    #   django-modeltranslation
    #   django-simple-menu
    #   django-user-agents
    #   djangorestframework
    #   djangosaml2
    #   sphinxcontrib-django
    #   uil-django-core
django-auth-ldap==4.6.0
    # via -r requirements.in
django-braces==1.15.0
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-impersonate==1.9.2
    # via -r requirements.in
django-modeltranslation==0.18.11
    # via -r requirements.in
django-simple-menu==2.1.3
    # via -r requirements.in
django-user-agents==0.4.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   uil-django-core
djangosaml2==1.6.0
    # via
    #   -r requirements.in
    #   cdh-django-core
djlint==1.34.1
    # via -r requirements.in
docutils==0.20.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
editorconfig==0.12.3
    # via
    #   cssbeautifier
    #   jsbeautifier
elementpath==4.1.5
    # via xmlschema
freetype-py==2.3.0
    # via
    #   reportlab
    #   rlpycairo
greenlet==3.0.3
    # via bpython
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
html5lib==1.1
    # via
    #   -r requirements.in
    #   xhtml2pdf
idna==3.6
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==7.0.1
    # via sphinx
jinja2==3.1.3
    # via sphinx
jsbeautifier==1.14.11
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.14
    # via djlint
lesscpy==0.15.1
    # via uil-django-core
lxml==5.1.0
    # via svglib
markupsafe==2.1.3
    # via jinja2
mypy-extensions==1.0.0
    # via black
mysqlclient==1.4.6
    # via -r requirements.in
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==23.2
    # via
    #   black
    #   sphinx
pathspec==0.12.1
    # via
    #   black
    #   djlint
pbr==6.0.0
    # via sphinxcontrib-apidoc
pdftotext==2.2.2
    # via -r requirements.in
pillow==10.2.0
    # via
    #   -r requirements.in
    #   reportlab
    #   xhtml2pdf
platformdirs==4.1.0
    # via black
ply==3.11
    # via lesscpy
pprintpp==0.4.0
    # via sphinxcontrib-django
pscript==0.7.7
    # via vbuild
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.3.0
    # via python-ldap
pycairo==1.25.1
    # via rlpycairo
pycparser==2.21
    # via cffi
pygments==2.17.2
    # via
    #   -r requirements.in
    #   bpython
    #   sphinx
pyhanko==0.21.0
    # via xhtml2pdf
pyhanko-certvalidator==0.26.3
    # via
    #   pyhanko
    #   xhtml2pdf
pyjwt==2.8.0
    # via uil-django-core
pyopenssl==23.3.0
    # via pysaml2
pypdf==3.17.4
    # via xhtml2pdf
pypng==0.20220715.0
    # via qrcode
pysaml2==7.4.2
    # via djangosaml2
pyscss==1.4.0
    # via uil-django-core
python-bidi==0.4.2
    # via xhtml2pdf
python-dateutil==2.8.2
    # via pysaml2
python-ldap==3.4.4
    # via django-auth-ldap
python-magic==0.4.27
    # via -r requirements.in
pytz==2023.3.post1
    # via
    #   django
    #   djangorestframework
    #   pysaml2
pyxdg==0.28
    # via bpython
pyyaml==6.0.1
    # via
    #   djlint
    #   pyhanko
qrcode==7.4.2
    # via pyhanko
regex==2023.12.25
    # via djlint
reportlab[pycairo]==4.0.9
    # via
    #   svglib
    #   xhtml2pdf
requests==2.31.0
    # via
    #   -r requirements.in
    #   bpython
    #   pyhanko
    #   pyhanko-certvalidator
    #   pysaml2
    #   sphinx
    #   uil-django-core
rlpycairo==0.3.0
    # via reportlab
six==1.16.0
    # via
    #   blessed
    #   cssbeautifier
    #   html5lib
    #   jsbeautifier
    #   pyscss
    #   python-bidi
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.2.6
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
    #   sphinxcontrib-apidoc
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-django
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-jquery
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
sphinx-rtd-theme==2.0.0
    # via -r requirements.in
sphinxcontrib-apidoc==0.4.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-django==2.5
    # via -r requirements.in
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
svglib==1.5.1
    # via xhtml2pdf
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
tomli==2.0.1
    # via
    #   black
    #   djlint
tqdm==4.66.1
    # via djlint
typing-extensions==4.9.0
    # via
    #   asgiref
    #   black
    #   django-modeltranslation
    #   pypdf
    #   qrcode
tzlocal==5.2
    # via pyhanko
ua-parser==0.18.0
    # via user-agents
uritools==4.0.2
    # via pyhanko-certvalidator
urllib3==2.1.0
    # via requests
user-agents==2.2.0
    # via django-user-agents
vbuild==0.8.2
    # via uil-django-core
wcwidth==0.2.13
    # via blessed
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wrapt==1.16.0
    # via deprecated
xhtml2pdf==0.2.13
    # via -r requirements.in
xmlschema==3.0.1
    # via pysaml2
zipp==3.17.0
    # via importlib-metadata
